# Optimize CI tooling with tools.go pattern