--- a/.github/workflows/checks.yml
+++ b/.github/workflows/checks.yml
@@ -19,9 +19,6 @@ jobs:
         with:
           go-version: ^1.24
 
-      - name: Download dependencies
-        run: go mod download
-
       - name: Run unit tests and generate the coverage report
         run: make test-race
 
@@ -37,17 +34,8 @@ jobs:
         with:
           go-version: ^1.24
 
-      - name: Download dependencies
-        run: go mod download
-
-      - name: Install gofumpt
-        run: go install mvdan.cc/gofumpt@v0.4.0
-
-      - name: Install staticcheck
-        run: go install honnef.co/go/tools/cmd/staticcheck@2025.1.1
-
-      - name: Install golangci-lint
-        run: go install github.com/golangci/golangci-lint/v2/cmd/golangci-lint@v2.1.2
+      - name: Install tools
+        run: make install-tools
 
       - name: Lint
         run: make lint