--- a/.gitignore
+++ b/.gitignore
@@ -19,3 +19,4 @@
 
 # Builds
 /build
+/bin/