--- a/tools/tools.go
+++ b/tools/tools.go
@@ -1,0 +1,9 @@
+//go:build tools
+
+package tools
+
+import (
+	_ "github.com/golangci/golangci-lint/v2/cmd/golangci-lint"
+	_ "honnef.co/go/tools/cmd/staticcheck"
+	_ "mvdan.cc/gofumpt"
+)