## Summary

Implements the idiomatic Go `tools.go` pattern to improve CI performance and developer experience.

## Performance Impact

**Before:** ~4 minutes for lint CI job  
**After:** ~90 seconds for lint CI job  
**Improvement:** 62% faster

### Why This Is Fast

The key optimization: **setup-go@v6's build cache** stores compiled object files (`.a` archives) in `~/.cache/go-build`. When we run `go install` for the same tool versions:

1. ✅ Modules already downloaded (cached in `~/go/pkg/mod`)
2. ✅ **Object files already compiled** (cached in `~/.cache/go-build`) ← **HUGE SAVINGS**
3. ❌ Only need to link the final binary (~5s)

**Result:** Tool installation goes from ~60s (cold compile) to ~10s (warm cache reuse).

Previously, the CI workflow hardcoded versions in YAML and installed tools fresh each run without leveraging setup-go@v6's caching. Now we use `tools.go` to track dependencies in `go.mod`, and setup-go@v6 automatically caches the expensive compilation step.

## Key Benefits

1. **62% faster CI** - Leverages setup-go@v6 build cache for compiled object files
2. **Single source of truth** - Tool versions tracked in `go.mod`, not hardcoded in YAML
3. **Project-local tools** - Each project has isolated tool versions in `./bin/` (no conflicts)
4. **Simpler workflow** - Just `make install-tools` locally or in CI

## Changes

- **tools/tools.go** - Tracks dev tool dependencies with build tag
- **Makefile** - Added `GOBIN=$(CURDIR)/bin` and `install-tools` target, updated lint targets
- **CI workflow** - Simplified to `make install-tools` (removed hardcoded versions)
- **.gitignore** - Added `/bin/` directory

## Tool Version Updates

| Tool | Before | After | Reason |
|------|--------|-------|--------|
| golangci-lint | v2.1.2 | v2.5.0 | v2.1.2 has broken dependencies in tools.go pattern |
| gofumpt | v0.4.0 | v0.9.1 | Required by golangci-lint v2.5.0 |
| staticcheck | 2025.1.1 | v0.6.1 | Latest stable |

## Testing

```bash
make install-tools  # Installs to ./bin/
make lint           # 0 issues
```

All linters pass cleanly with the updated versions.

## Pattern Reference

This follows the standard Go tooling pattern documented in the community research and used by major projects.